Add "lean" versions of summary functions
These funcions do not require the "Name" columns but
only the corresponding "ID" columns. The merge of the
"Name" columns is done after summarising. This allows
for smaller input data frames and much faster
processing.

Add "reverse_engineer.R", a script to test whether
the exact same results as with the original
functions can be achieved with the "lean" functions,
called on the "lean" data frames of the risk object.